Update namespace for filters
