--- api/app/presenters/spree/products/filters_presenter.rb
@@ -1,35 +1,0 @@
-module Spree
-  module Products
-    class FiltersPresenter
-      def initialize(current_store, current_currency, params)
-        @products_for_filters = find_products_for_filters(current_store, current_currency, params)
-      end
-
-      def to_h
-        option_values = Spree::OptionValues::FindAvailable.new(products_scope: products_for_filters).execute
-        product_properties = Spree::ProductProperties::FindAvailable.new(products_scope: products_for_filters).execute
-
-        {
-          option_types: Spree::Products::OptionTypeFiltersPresenter.new(option_values).to_a,
-          product_properties: Spree::Products::PropertyFiltersPresenter.new(product_properties).to_a
-        }
-      end
-
-      private
-
-      attr_reader :products_for_filters
-
-      def find_products_for_filters(current_store, current_currency, params)
-        current_taxons = find_current_taxon(current_store, params)
-        current_store.products.active(current_currency).in_taxons(current_taxons)
-      end
-
-      def find_current_taxon(current_store, params)
-        taxons_param = params.dig(:filter, :taxons)
-        return nil if taxons_param.nil? || taxons_param.to_s.blank?
-
-        current_store.taxons.find_by(id: taxons_param)
-      end
-    end
-  end
-end
